splitted SnippetMenu to externalise the modale section
